% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dash_bar}
\alias{dash_bar}
\title{Navbar with no contents}
\usage{
dash_bar(..., label = NULL)
}
\arguments{
\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Navbar labels}
}
\description{
Navbar with no contents
}
