% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dash_input}
\alias{dash_input}
\title{Create inputs}
\usage{
dash_input(..., .envir = parent.frame())
}
\arguments{
\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Input specs}
}
\description{
Create inputs
}
