% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{eval_env}
\alias{eval_env}
\title{Get environment (list) of input/calc variables for a given output/calc}
\usage{
eval_env(d, input, req_input, calc, req_calc)
}
\description{
Get environment (list) of input/calc variables for a given output/calc
}
