% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Highlight areas on a plot}
\usage{
highlight(
  x = NULL,
  y = NULL,
  col = "#ffff00",
  label = NA,
  alpha = 0.2,
  shrink = TRUE,
  border = NA,
  ...
)
}
\arguments{
\item{x}{x coordinates for highlight areas.}

\item{y}{y coordinates for highlight areas (usually leave \code{NULL}).}

\item{col}{fill colours for rectangles.}

\item{label}{text labels for each rectangle.}

\item{alpha}{opacity level (0 to 1) for rectangles.}

\item{shrink}{if \code{TRUE} or numeric and \code{y} is \code{NULL}, lowers
the top of each subsequent rectangle by the specified number of user
plotting units so that the labels don't overlap; \code{TRUE} uses the
line height times 2, which is a decent default. Set to zero for no
shrinking.}

\item{border}{colour for rectangle borders; \code{NA} (default) to omit.}

\item{...}{other arguments to pass on to \code{rect()}.}
}
\description{
Draws a series of highlights (translucent rectangles) onto a base R plot.
}
\details{
\code{x} and \code{y} can either be: \code{NULL}, in which case the extent
of the plotted area is used; a numeric vector, interleaved such that the
elements are interpreted as e.g.
\code{c(x1_start, x1_end, x2_start, x2_end, ...)}; a list with two elements,
e.g. \code{list(c(x1_start, x2_start, ...), c(x1_end, x2_end, ...))}; or a
matrix with two columns (first column "starts", second column "ends").

Arguments are recycled as needed.
}
\examples{
\dontrun{
curve(dnorm(x), 0, 4)
highlight(x = c(0, 0, 1, 3, 2, 4), label = letters[1:3], col = c("red", "blue", "purple"))
}
}
